+++ asyncgit/src/sync/branch/mod.rs
@@ -11,8 +11,10 @@ use crate::{
 	sync::{
 		remotes::get_default_remote_for_push_in_repo,
 		repository::repo, utils::get_head_repo, CommitId,
+		CommitSignature,
 	},
 };
+use chrono::{DateTime, Local, TimeZone};
 use git2::{Branch, BranchType, Repository};
 use scopetime::scope_time;
 use std::collections::HashSet;
@@ -91,6 +93,12 @@ pub struct BranchInfo {
 	pub top_commit_message: String,
 	///
 	pub top_commit: CommitId,
+	///
+	pub top_commit_time: i64,
+	///
+	pub top_commit_time_local: Option<DateTime<Local>>,
+	///
+	pub top_commit_author: String,
 	///
 	pub details: BranchDetails,
 }
@@ -181,6 +189,8 @@ pub fn get_branches_info(
 				})
 			};
 
+			let author = CommitSignature::from(&top_commit.author());
+
 			Ok(BranchInfo {
 				name: bytes2string(name_bytes)?,
 				reference,
@@ -188,6 +198,11 @@ pub fn get_branches_info(
 					top_commit.summary_bytes().unwrap_or_default(),
 				)?,
 				top_commit: top_commit.id().into(),
+				top_commit_time: top_commit.time().seconds(),
+				top_commit_time_local: Local
+					.timestamp_opt(top_commit.time().seconds(), 0)
+					.earliest(),
+				top_commit_author: author.name,
 				details,
 			})
 		})