+++ src/popups/mod.rs
@@ -1,4 +1,5 @@
 mod blame_file;
+mod branch_sort;
 mod branchlist;
 mod commit;
 mod compare_commits;
@@ -25,6 +26,7 @@ mod tag_commit;
 mod taglist;
 
 pub use blame_file::{BlameFileOpen, BlameFilePopup};
+pub use branch_sort::BranchSortPopup;
 pub use branchlist::BranchListPopup;
 pub use commit::CommitPopup;
 pub use compare_commits::CompareCommitsPopup;